Creator/requirements.txt
cryptography
Flask
multiversx-sdk-core
multiversx-sdk-wallet
multiversx_sdk_network_providers
multiversx_sdk
Pillow
transformers
diffusers
ipfshttpclient
sentencepiece